ui: improves file preview on reply
