+++ src/lib/Room/RoomFooter/RoomMessageReply/RoomMessageReply.vue
@@ -39,19 +39,19 @@
         </audio-player>
 
         <div v-else-if="isOtherFile" class="vac-image-reply vac-file-container">
-          <div>
-            <slot name="file-icon">
-              <svg-icon name="file" />
-            </slot>
+          <div class="vac-reply-file-icon">
+            <i :class="fileIconClass" style="font-size: 35px;" />
           </div>
-          <div class="vac-text-ellipsis">
-            {{ firstFile.name }}
-          </div>
-          <div
-            v-if="firstFile.extension"
-            class="vac-text-ellipsis vac-text-extension"
-          >
-            {{ firstFile.extension }}
+          <div class="vac-reply-file-info">
+            <div class="vac-text-ellipsis vac-reply-file-name">
+              {{ firstFile.name }}
+            </div>
+            <div
+              v-if="firstFile.extension"
+              class="vac-reply-file-extension-and-size"
+            >
+              {{ fileExtensionAndSize }}
+            </div>
           </div>
         </div>
       </div>
@@ -72,6 +72,8 @@ import SvgIcon from '../../../../components/SvgIcon/SvgIcon'
 import FormatMessage from '../../../../components/FormatMessage/FormatMessage'
 
 import AudioPlayer from '../../RoomMessage/AudioPlayer/AudioPlayer'
+
+import { humanFileSize } from '../../../../utils/adhoc'
 
 import {
   isAudioFile,
@@ -116,6 +118,12 @@ export default {
         !this.isVideo &&
         !this.isImage
       )
+    },
+    fileIconClass() {
+      return Optidata.MimeTypeIcons.getIconByMimeType(this.firstFile.type)
+    },
+    fileExtensionAndSize() {
+      return `${this.firstFile.extension} · ${humanFileSize(this.firstFile.size, true)}`
     }
   }
 }
